requirements.txt
Flask==1.1.2
line-bot-sdk==1.17.0
requests==2.24.0